Video Background: transparent
Made background color of video transparent, instead of black. Untested, I didn't take time to see if the css 'cascades' properly. There is a chance it doesn't override the background. Also I don't care what the color is, I just wanted to be able to see the outlines of angry snoo.